backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
supabase==2.9.0
google-generativeai>=0.8.3
requests==2.32.3
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator>=2.0.0
dateutils==0.6.12
apscheduler==3.10.4

